+++ include/redirections.h
@@ -27,7 +27,7 @@ bool open_pipe_redirections(command_t *command);
 void handle_pipe_redirections(command_t *command);
 
 bool check_redirections(command_t *list, minishell_t *shell, varenv_t *env);
-redirection_t get_redirection(char *str, int index);
+redirection_t get_redirection(char *str, size_t index);
 char **split_redirections(char *input);
 
 void replace_args(command_t *command, char *redir, char *str, varenv_t *env);
